// js/stat-evaluation-player/src/generated/HalfVolleyPlayerStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Per-player accumulated half-volley stats.
 */
export type HalfVolleyPlayerStats = { count: number, total_ball_speed: number, fastest_ball_speed: number, is_last_half_volley: boolean, last_half_volley_time: number | null, last_half_volley_frame: number | null, time_since_last_half_volley: number | null, frames_since_last_half_volley: number | null, };